training/pyproject.toml
[project]
name = "deni-train"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "OpenEXR>=3.2",
    "Imath>=0.0.1",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "tensorboard>=2.12",
    "Pillow>=10.0",
    "onnx>=1.14",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "scipy>=1.10"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
